add role object for readability, refactor roleMiddleware functionality
