--- api/models/role.js
@@ -1,8 +1,0 @@
-class Role {
-    constructor(roleId, role) {
-        this.roleId = roleId;
-        this.role = role;
-    }
-}
-
-module.exports = Role;
