+++ api/util/role.js
@@ -1,0 +1,8 @@
+
+const Role = Object.freeze({
+    ADMIN: 1,
+    TEACHER: 2,
+    STUDENT: 3
+})
+
+module.exports = Role;
